requirements.txt
Django>=4.2,<5.0
python-decouple>=3.8
psycopg2-binary>=2.9
djangorestframework>=3.14
dj-database-url>=2.0
pytest>=7.0
pytest-django>=4.5